backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
dspy>=2.6.27
faster-whisper>=1.1.1
torch>=2.7.1
mlflow>=3.1.0
openai-whisper @ git+https://github.com/openai/whisper.git
pydub>=0.25.1
pyvis>=0.3.2
python-multipart